Move cffi bindings to own module
