+++ numba/cuda/target.py
@@ -3,6 +3,7 @@ from functools import cached_property
 import llvmlite.binding as ll
 from llvmlite import ir
 
+from numba.cffi import cffi_utils
 from numba.core import (cgutils, config, debuginfo, itanium_mangler, types,
                         typing, utils)
 from numba.core.dispatcher import Dispatcher
@@ -22,7 +23,7 @@ from numba.cuda.models import cuda_data_manager
 class CUDATypingContext(typing.BaseContext):
     def load_additional_registries(self):
         from . import cudadecl, cudamath, libdevicedecl, vector_types
-        from numba.core.typing import enumdecl, cffi_utils
+        from numba.core.typing import enumdecl
 
         self.install_registry(cudadecl.registry)
         self.install_registry(cffi_utils.registry)
