+++ numba/tests/cffi_usecases.py
@@ -2,7 +2,7 @@ import sys
 
 import numpy as np
 
-import numba.core.typing.cffi_utils as cffi_support
+import numba.cffi.cffi_utils as cffi_support
 from numba.tests.support import import_dynamic, temp_directory
 from numba.core.types import complex128
