+++ numba/tests/test_cffi.py
@@ -2,7 +2,7 @@ import array
 import numpy as np
 
 from numba import jit, njit
-import numba.core.typing.cffi_utils as cffi_support
+import numba.cffi.cffi_utils as cffi_support
 from numba.core import types, errors
 from numba.tests.support import TestCase, skip_unless_cffi
